backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
databases
python-dotenv
pydantic
pydantic-settings
types-python-dateutil
openai
anthropic
redis
celery
requests
httpx
sentence-transformers
pgvector
passlib[bcrypt]
python-jose[cryptography]
bcrypt